rewrite EndpointTCPServer, EndpointUDPServer, EndpointTCPClient, EndpointUDPClient
... as children of EndpointCustomServer and EndpointCustomClient
