+++ endpoint_client.go
@@ -3,7 +3,6 @@ package gomavlib
 import (
 	"context"
 	"io"
-	"net"
 	"time"
 
 	"github.com/bluenviron/gomavlib/v3/pkg/timednetconn"
@@ -11,65 +10,9 @@ import (
 
 var reconnectPeriod = 2 * time.Second
 
-type endpointClientConf interface {
-	init(*Node) (Endpoint, error)
-}
-
-// EndpointTCPClient sets up a endpoint that works with a TCP client.
-// TCP is fit for routing frames through the internet, but is not the most
-// appropriate way for transferring frames from a UAV to a GCS, since it does
-// not allow frame losses.
-type EndpointTCPClient struct {
-	// domain name or IP of the server to connect to, example: 1.2.3.4:5600
-	Address string
-}
-
-func (conf EndpointTCPClient) init(node *Node) (Endpoint, error) {
-	e := &endpointClient{
-		node: node,
-		conf: conf,
-	}
-	err := e.initialize()
-	return e, err
-}
-
-// EndpointUDPClient sets up a endpoint that works with a UDP client.
-type EndpointUDPClient struct {
-	// domain name or IP of the server to connect to, example: 1.2.3.4:5600
-	Address string
-}
-
-func (conf EndpointUDPClient) init(node *Node) (Endpoint, error) {
-	e := &endpointClient{
-		node: node,
-		conf: conf,
-	}
-	err := e.initialize()
-	return e, err
-}
-
-// EndpointCustomClient sets up a endpoint that works with a custom implementation
-// by providing a Connect func that returns a net.Conn.
-type EndpointCustomClient struct {
-	// custom connect function that opens the connection
-	Connect func(ctx context.Context) (net.Conn, error)
-
-	// the label of the protocol
-	Label string
-}
-
-func (conf EndpointCustomClient) init(node *Node) (Endpoint, error) {
-	e := &endpointClient{
-		node: node,
-		conf: conf,
-	}
-	err := e.initialize()
-	return e, err
-}
-
 type endpointClient struct {
 	node *Node
-	conf endpointClientConf
+	conf EndpointCustomClient
 
 	ctx       context.Context
 	ctxCancel func()
@@ -97,24 +40,8 @@ func (e *endpointClient) oneChannelAtAtime() bool {
 }
 
 func (e *endpointClient) connect() (io.ReadWriteCloser, error) {
-	// in UDP, the only possible error is a DNS failure
-	// in TCP, the handshake must be completed
 	timedContext, timedContextClose := context.WithTimeout(e.ctx, e.node.ReadTimeout)
-	nconn, err := func() (net.Conn, error) {
-		switch conf := e.conf.(type) {
-		case EndpointTCPClient:
-			return (&net.Dialer{}).DialContext(timedContext, "tcp4", conf.Address)
-
-		case EndpointUDPClient:
-			return (&net.Dialer{}).DialContext(timedContext, "udp4", conf.Address)
-
-		case EndpointCustomClient:
-			return conf.Connect(timedContext)
-
-		default:
-			panic("should not happen")
-		}
-	}()
+	nconn, err := e.conf.Connect(timedContext)
 	timedContextClose()
 
 	if err != nil {
@@ -155,17 +82,8 @@ func (e *endpointClient) provide() (string, io.ReadWriteCloser, error) {
 }
 
 func (e *endpointClient) label() string {
-	switch conf := e.conf.(type) {
-	case EndpointTCPClient:
-		return "tcp:" + conf.Address
-	case EndpointUDPClient:
-		return "udp:" + conf.Address
-	case EndpointCustomClient:
-		if conf.Label != "" {
-			return conf.Label
-		}
-		return "custom"
-	default:
-		return "unknown"
+	if e.conf.Label != "" {
+		return e.conf.Label
 	}
+	return "custom"
 }
